programming exercise stellar evolution 2021
In the theory of stellar structure and stellar evolution the objects, i.e. the stars, can be described by the evolution of the density, the mass, the temperature, the pressure, and the composition over its radius. Some of the equations describing these elements get quite complicated as they are represented by differential equations. The aim of the computer exercise is to take on a pair of coupled differential equations numerically, and thereby describing the structure of a white dwarf.